Beginning of a consolidated login flow: WIP
* Replace Login dropdown with a simple Login button
* Introduce a page that captures just the user email
* Based the email, render either the partner or bank login page, with
email pre-filled
* Introduce a new system spec to cover this feature

The Users::LookupsController does not derive from Devise, mostly because
i couldn't get that to work. Instead it introduces some helper methods
that essential duck type it so it looks like a devise controller. This
allows it to directly render devise templates instead of redirecting to
their respective controller.